flask backend/requirement.txt
Flask
Flask-cors
Flask_socketio
PyExecJS
Flask-PyMongo